- split IT tests
